server/requirements.txt
flask==3.0.0
numpy==1.26.2
pillow==10.1.0
tensorflow==2.15.0
matplotlib==3.8.2
flask-cors==4.0.0
scikit-learn==1.3.2
seaborn==0.13.0
gunicorn==21.2.0
joblib==1.3.2
